requirements.txt
streamlit==1.44.1
openai>=1.30.0
python-dotenv>=1.0.1
pydantic>=2.7.0